make consumable via npm
- add linting
