+++ .gitignore
@@ -1,2 +1,1 @@
-/bower_components
 /node_modules
