--- bower.json
@@ -1,14 +1,0 @@
-{
-  "name": "knockout-date-picker",
-  "version": "2.0.0",
-  "main": "knockout.date-picker.js",
-  "ignore": [
-    "node_modules",
-    "bower_components"
-  ],
-  "dependencies": {
-    "knockout": "~3.1.0",
-    "jquery": "~1.10",
-    "jquery.ui": "~1.10.4"
-  }
-}
